BUG: Python3: Review and fix use of division
In Python3, the division is a "true division" and not a "floor division"

See https://www.python.org/dev/peps/pep-0238/
PEP 238 -- Changing the Division Operator

This commit was crafted applying the following steps:

- installing future

  Slicer_DIR=/path/to/Slicer-SuperBuild/Slicer-build
  ${Slicer_DIR}/../python-install/bin/PythonSlicer -m pip install
future

- applying the "libfuturize.fixes.fix_division_safe" fix

  for f in `find ./ -name *.py`; do \
    ${Slicer_DIR}/../python-install/bin/PythonSlicer \
      --launch futurize -f libfuturize.fixes.fix_division_safe -w $f; \
  done

- reverting addition of "from __future__ import division"
  and "from past.utils import old_div"

  for f in `find ./ -name "*.py"`; do \
    sed -i '/from __future__ import division/ d' $f; \
    sed -i '/from past.utils import old_div/ d' $f; \
  done

- removing use of "old_div" function and replace with use of int() where
  relevant. This is step is *NOT* automated.

svn-url: http://viewvc.slicer.org/viewvc.cgi/Slicer4?view=revision&revision=28110
git-svn-id: http://svn.slicer.org/Slicer4/trunk@28110 3bd1e089-480b-0410-8dfb-8563597acbee

+++ Applications/SlicerApp/Testing/Python/sceneImport2428.py
@@ -127,8 +127,8 @@ class sceneImport2428Test(ScriptedLoadableModuleTest):
     self.delayDisplay('Paint radius is %s' % parameterNode.GetParameter('PaintEffect,radius'))
     sliceWidget = lm.sliceWidget('Red')
     size = min(sliceWidget.width,sliceWidget.height)
-    step = size / 12
-    center = size / 2
+    step = int(size / 12)
+    center = int(size / 2)
     parameterNode.SetParameter('PaintEffect,radius', '20')
     paintTool = EditorLib.PaintEffectTool(sliceWidget)
     self.delayDisplay('Paint radius is %s, tool radius is %d' % (parameterNode.GetParameter('PaintEffect,radius'),paintTool.radius))
